moving classes into several files, and making non-private useful functions
