+++ tests/u2flib_test.php
@@ -66,7 +66,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_ATTESTATION_VERIFICATION
    */
   public function testDoRegisterAttestFail() {
@@ -77,7 +77,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_ATTESTATION_SIGNATURE
    */
   public function testDoRegisterFail2() {
@@ -87,7 +87,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_UNMATCHED_CHALLENGE
    */
   public function testDoRegisterFail() {
@@ -107,7 +107,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_PUBKEY_DECODE
    */
   public function testDoRegisterBadKeyInCert() {
@@ -117,7 +117,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_PUBKEY_DECODE
    */
   public function testDoRegisterBadKey() {
@@ -147,7 +147,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_BAD_UA_RETURNING
    */
   public function testDoRegisterUAError() {
@@ -194,7 +194,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_COUNTER_TOO_LOW
    */
   public function testDoAuthenticateCtrFail() {
@@ -205,7 +205,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_AUTHENTICATION_FAILURE
    */
   public function testDoAuthenticateFail() {
@@ -216,7 +216,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_NO_MATCHING_REQUEST
    */
   public function testDoAuthenticateWrongReq() {
@@ -227,7 +227,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_NO_MATCHING_REGISTRATION
    */
   public function testDoAuthenticateWrongReg() {
@@ -238,7 +238,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_PUBKEY_DECODE
    */
   public function testDoAuthenticateBadKey() {
@@ -282,7 +282,7 @@ class U2FTest extends \PHPUnit_Framework_TestCase {
   }
 
   /**
-   * @expectedException u2flib_server\Error
+   * @expectedException \u2flib_server\Error
    * @expectedExceptionCode u2flib_server\ERR_BAD_UA_RETURNING
    */
   public function testDoAuthenticateUAError() {
